+++ templates/rvo/detail.html
@@ -1,4 +1,4 @@
-{% extends "base-rvo.html" %}
+{% extends "rvo/base.html" %}
 
 {% block footer_nav_current %}rvo{% endblock %}
